requirements.txt
anthropic>=0.25.0
azure-ai-ml>=1.13.0
azure-monitor-query>=1.3.0
azure-identity>=1.15.0
jira>=3.6.0
schedule>=1.2.1
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.3
